Refactor ManagedApplication into ManagedDeployment and ManagedStatefulSet to improve readibility
